refactor(model): move Task data model for better UI structure
